Split monomorphization dependency overlays
