+++ docs/PHASE_PLAN.md
@@ -198,6 +198,9 @@ checked-in docs, tests, and commits only.
   module, keeping the public lexer API and character-span helpers compact.
 - Parser block, closure, and argument-list helpers now live in a dedicated
   parser module, keeping atom parsing focused on prefix expression dispatch.
+- Typechecker monomorphization template dependency install/restore now lives in
+  a dedicated helper module, keeping generic specialization flow separate from
+  temporary source-module dependency overlays.
 - Module-system unit tests now live beside the module-system implementation in
   a dedicated test module, keeping module graph/load entry points compact.
 - Resolver expected-local traversal now lives in a dedicated validation support
