+++ src/typechecker/mod.rs
@@ -21,6 +21,7 @@ mod generic_bound_validation;
 mod generic_type_reference_walker;
 mod generic_type_validation;
 mod monomorphize;
+mod monomorphize_dependencies;
 mod monomorphize_inference;
 mod monomorphize_types;
 mod patterns;
